backend/app/requirements/development.txt
fastapi
uvicorn
debugpy
pymongo
pyjwt
pydantic[email]